many features added

+++ routes/web.php
@@ -1,8 +1,11 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CommentController;
 use App\Http\Controllers\DashboardController;
+use App\Http\Controllers\FollowerController;
 use App\Http\Controllers\IdeaController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -18,22 +21,75 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/', [DashboardController::class, 'index'])->name('dashboard');
 
+/*
+|
+|
+|--------------------------------------------------------------------------
+| Group Routes
+|--------------------------------------------------------------------------
+|
+|
+*/
+
+Route::group(['prefix'=>'ideas/','as'=>'ideas.'], function(){
+    Route::post('', [IdeaController::class, 'store'])->name('store');
+
+    Route::get('/{idea}', [IdeaController::class, 'show'])->name('show');
+
+    Route::group(['middleware' => ['auth']], function () {
+
+        Route::get('/{idea}/edit', [IdeaController::class, 'edit'])->name('edit');
+
+        Route::put('/{idea}', [IdeaController::class, 'update'])->name('update');
+
+        Route::delete('/{idea}', [IdeaController::class, 'destroy'])->name('destroy');
+
+        Route::post('/{idea}/comments', [CommentController::class, 'store'])->name('comments.store');
+
+    });
+
+});
+
+/*
 Route::post('/ideas', [IdeaController::class, 'store'])->name('ideas.store');
 
 Route::get('/ideas/{idea}', [IdeaController::class, 'show'])->name('ideas.show');
 
-Route::get('/ideas/{idea}/edit', [IdeaController::class, 'edit'])->name('ideas.edit');
+Route::get('/ideas/{idea}/edit', [IdeaController::class, 'edit'])->name('ideas.edit')->middleware('auth');
 
-Route::put('/ideas/{idea}', [IdeaController::class, 'update'])->name('ideas.update');
+Route::put('/ideas/{idea}', [IdeaController::class, 'update'])->name('ideas.update')->middleware('auth');
 
-Route::delete('/ideas/{idea}', [IdeaController::class, 'destroy'])->name('ideas.destroy');
+Route::delete('/ideas/{idea}', [IdeaController::class, 'destroy'])->name('ideas.destroy')->middleware('auth');
 
-Route::post('/ideas/{idea}/comments', [CommentController::class, 'store'])->name('ideas.comments.store');
-
+Route::post('/ideas/{idea}/comments', [CommentController::class, 'store'])->name('ideas.comments.store')->middleware('auth');
+ */
 Route::get('/terms', function () {
     return view('terms');
 });
 
 Route::get('/register', [AuthController::class, 'register'])->name('register');
 
-Route::get('/register', [AuthController::class, 'store']);
+Route::post('/register', [AuthController::class, 'store']);
+
+Route::get('/login', [AuthController::class, 'login'])->name('login');
+
+Route::post('/login', [AuthController::class, 'authenticate']);
+
+Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
+
+/*
+|
+|
+|--------------------------------------------------------------------------
+| Resource Routes of users(profile)
+|--------------------------------------------------------------------------
+|
+|
+*/
+Route::resource('users', UserController::class)->only('show', 'edit', 'update')->middleware('auth');
+
+Route::get('profile', [UserController::class, 'profile'])->middleware('auth')->name('profile');
+
+Route::post('users/{user}/follow', [FollowerController::class, 'follow'])->middleware('auth')->name('users.follow');
+
+Route::post('users/{user}/unfollow', [FollowerController::class, 'unfollow'])->middleware('auth')->name('users.unfollow');
